+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## master branch
 
   - add `optional_tag` setting so users can keep ", optional" from showing up by default
+  - support multiple parameters with the same description for numpy style
 
 ## 0.2.5
   Features: